Create SankeWaterGun
It is a game that van  be played with computer ,it's similar to stone paper scissor . In this you need choose one among the snake, water and gun randomly and the  computer also does so. Snake can drink the water, gun can kill snake and the water can malfunction the gun so for example if you choose snake and opponent choose gun then opponent wins but if opponent chooses water then you win and this is how it is played. the code is python hope you enjoy playing it.